Подставляет заглушку для отсутствия задач и рефакторит вспомогательные функции
