+++ src/view/filter-view.js
@@ -5,7 +5,7 @@ function createFilterTemplate() {
     `<form class="trip-filters" action="#" method="get">
       <div class="trip-filters__filter">
         <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter"
-          value="everything">
+          value="everything" checked>
         <label class="trip-filters__filter-label" for="filter-everything">Everything</label>
       </div>
 
@@ -23,7 +23,7 @@ function createFilterTemplate() {
 
       <div class="trip-filters__filter">
         <input id="filter-past" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter"
-          value="past" checked>
+          value="past">
         <label class="trip-filters__filter-label" for="filter-past">Past</label>
       </div>
