+++ src/view/no-event-view.js
@@ -1,0 +1,20 @@
+import AbstractView from '../framework/view/abstract-view.js';
+
+function createNoEventTemplate() {
+  return `
+    <p class="trip-events__msg">
+      Click New Event to create your first point
+    </p>
+  `;
+}
+
+// Everything – 'Click New Event to create your first point'
+// Past — 'There are no past events now';
+// Present — 'There are no present events now';
+// Future — 'There are no future events now'.
+
+export default class NoEventView extends AbstractView {
+  get template() {
+    return createNoEventTemplate();
+  }
+}
